update js docs
