+++ login-workflow/src/contexts/ErrorContext/types.ts
@@ -3,6 +3,6 @@
  * @module ErrorContext
  */
 
-import { ErrorManagerProps } from '../../components/Error/ErrorManager';
+import { ErrorManagerProps } from '../../components/Error/types';
 
 export type ErrorContextProviderProps = Omit<ErrorManagerProps, 'error'>;
