+++ login-workflow/src/screens/AccountDetailsScreen/types.ts
@@ -1,6 +1,6 @@
 import { BoxProps, TextFieldProps } from '@mui/material';
 import { WorkflowCardProps } from '../../components/WorkflowCard/WorkflowCard.types';
-import { ErrorManagerProps } from '../../components/Error';
+import { ErrorManagerProps } from '../../components/Error/types';
 
 export type AccountDetailsScreenProps = WorkflowCardProps &
     BoxProps & {
