+++ login-workflow/src/screens/CreateAccountScreen/types.ts
@@ -1,6 +1,6 @@
 import { TextFieldProps } from '@mui/material';
 import { WorkflowCardProps } from '../../components/WorkflowCard/WorkflowCard.types';
-import { ErrorManagerProps } from '../../components/Error';
+import { ErrorManagerProps } from '../../components/Error/types';
 
 export type CreateAccountScreenProps = WorkflowCardProps & {
     /**
