+++ login-workflow/src/screens/CreatePasswordScreen/types.ts
@@ -1,5 +1,5 @@
 import { SetPasswordProps } from '../../components';
-import { ErrorManagerProps } from '../../components/Error';
+import { ErrorManagerProps } from '../../components/Error/types';
 import { WorkflowCardProps } from '../../components/WorkflowCard/WorkflowCard.types';
 
 export type CreatePasswordScreenProps = WorkflowCardProps & {
