+++ login-workflow/src/screens/ResetPasswordScreen/types.ts
@@ -1,4 +1,4 @@
-import { ErrorManagerProps } from '../../components/Error';
+import { ErrorManagerProps } from '../../components/Error/types';
 import { SetPasswordProps } from '../../components/SetPassword/types';
 import { WorkflowCardProps } from '../../components/WorkflowCard/WorkflowCard.types';
 import { SuccessScreenProps } from '../SuccessScreen';
