+++ login-workflow/src/screens/VerifyCodeScreen/types.ts
@@ -1,5 +1,5 @@
 import { TextFieldProps } from '@mui/material';
-import { ErrorManagerProps } from '../../components/Error';
+import { ErrorManagerProps } from '../../components/Error/types';
 import { WorkflowCardProps } from '../../components/WorkflowCard/WorkflowCard.types';
 
 export type VerifyCodeScreenProps = WorkflowCardProps & {
